Settings: Fix devloper options FC and remove all bug report options
This change: https://gerrit.slimroms.net/#/c/18820/

removed it just in the xml which is a pretty bad idea if java code
still assumes it is there.

Due that we will never need it we remove the code completely.

As well remove bug report in power menu option.

Change-Id: I234dee3a19639f89b5a7b5295ca408142e4223dc
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>

Conflicts:
	res/xml/development_prefs.xml
	src/com/android/settings/DevelopmentSettings.java

Conflicts:
	res/xml/development_prefs.xml
	src/com/android/settings/DevelopmentSettings.java

Signed-off-by: Chet Kener <Cl3Kener@gmail.com>

Conflicts:
	src/com/android/settings/DevelopmentSettings.java